# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Bingo API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TelegramAuthInput {
  initData: string;
}

export interface AdminLoginInput {
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  player: 'player',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  telegramId: string;
  username: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  balance: number;
  totalWinnings?: number;
  totalGames?: number;
  totalWins?: number;
  role: UserRole;
  isBanned: boolean;
  referralCode?: string;
  /** @nullable */
  referredBy?: number | null;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  user: User;
}

export interface UserPage {
  data: User[];
  total: number;
  page: number;
  limit: number;
}

export interface BalanceAdjustInput {
  amount: number;
  reason: string;
}

export interface BanUserInput {
  banned: boolean;
}

export interface UserStats {
  totalGames: number;
  totalWins: number;
  totalWinnings: number;
  totalSpent: number;
  winRate: number;
}

export type GameStatus = typeof GameStatus[keyof typeof GameStatus];


export const GameStatus = {
  waiting: 'waiting',
  active: 'active',
  finished: 'finished',
} as const;

export interface Game {
  id: number;
  status: GameStatus;
  ticketPrice: number;
  totalPot: number;
  houseFee: number;
  prizePool: number;
  houseEarnings: number;
  drawnNumbers: number[];
  playerCount: number;
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerUsername?: string | null;
  /** @nullable */
  winPattern?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
  createdAt: string;
}

export interface GamePage {
  data: Game[];
  total: number;
  page: number;
  limit: number;
}

export interface Ticket {
  id: number;
  gameId: number;
  userId: number;
  card: number[][];
  markedNumbers: number[];
  isWinner: boolean;
  /** @nullable */
  prizeAmount?: number | null;
  purchasedAt: string;
}

export interface TicketPage {
  data: Ticket[];
  total: number;
  page: number;
  limit: number;
}

export interface BingoClaimInput {
  ticketId: number;
}

export interface BingoClaimResult {
  valid: boolean;
  message: string;
  /** @nullable */
  prizeAmount?: number | null;
  /** @nullable */
  pattern?: string | null;
}

export type DepositStatus = typeof DepositStatus[keyof typeof DepositStatus];


export const DepositStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Deposit {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  amount: number;
  status: DepositStatus;
  telebirrNumber: string;
  transactionId: string;
  /** @nullable */
  screenshotUrl?: string | null;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
}

export interface DepositPage {
  data: Deposit[];
  total: number;
  page: number;
  limit: number;
}

export interface DepositInput {
  /** @minimum 1 */
  amount: number;
  telebirrNumber: string;
  transactionId: string;
  screenshotUrl?: string;
}

export type WithdrawalStatus = typeof WithdrawalStatus[keyof typeof WithdrawalStatus];


export const WithdrawalStatus = {
  pending: 'pending',
  approved: 'approved',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export interface Withdrawal {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  amount: number;
  status: WithdrawalStatus;
  telebirrNumber: string;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
}

export interface WithdrawalPage {
  data: Withdrawal[];
  total: number;
  page: number;
  limit: number;
}

export interface WithdrawalInput {
  /** @minimum 1 */
  amount: number;
  telebirrNumber: string;
}

export interface RejectInput {
  reason: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  firstName?: string;
  /** @nullable */
  avatarUrl?: string | null;
  totalWinnings: number;
  totalWins: number;
}

export interface HouseSettings {
  id: number;
  /** House fee percentage (default 30) */
  houseFeePct: number;
  ticketPrice: number;
  /** Milliseconds between number draws */
  drawIntervalMs: number;
  /** Lobby countdown seconds */
  countdownSecs: number;
  minPlayers: number;
  maxPlayers: number;
  adminTelebirrNumber?: string;
  updatedAt: string;
}

export interface HouseSettingsUpdate {
  /**
     * @minimum 0
     * @maximum 100
     */
  houseFeePct?: number;
  /** @minimum 1 */
  ticketPrice?: number;
  /** @minimum 1000 */
  drawIntervalMs?: number;
  /** @minimum 10 */
  countdownSecs?: number;
  /** @minimum 1 */
  minPlayers?: number;
  maxPlayers?: number;
  adminTelebirrNumber?: string;
}

export interface Announcement {
  id: number;
  title: string;
  message: string;
  createdAt: string;
}

export interface AnnouncementInput {
  title: string;
  message: string;
}

export interface DashboardStats {
  totalRevenue: number;
  houseEarnings: number;
  prizePoolTotal: number;
  dailyProfit: number;
  weeklyProfit: number;
  monthlyProfit: number;
  totalTicketsSold: number;
  avgPlayersPerGame: number;
  totalUsers: number;
  activeGames: number;
  pendingDeposits: number;
  pendingWithdrawals: number;
}

export type ListUsersParams = {
page?: number;
limit?: number;
search?: string;
};

export type ListGamesParams = {
page?: number;
limit?: number;
status?: ListGamesStatus;
};

export type ListGamesStatus = typeof ListGamesStatus[keyof typeof ListGamesStatus];


export const ListGamesStatus = {
  waiting: 'waiting',
  active: 'active',
  finished: 'finished',
} as const;

export type ListMyTicketsParams = {
page?: number;
limit?: number;
};

export type ListDepositsParams = {
page?: number;
limit?: number;
status?: ListDepositsStatus;
};

export type ListDepositsStatus = typeof ListDepositsStatus[keyof typeof ListDepositsStatus];


export const ListDepositsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type ListWithdrawalsParams = {
page?: number;
limit?: number;
status?: ListWithdrawalsStatus;
};

export type ListWithdrawalsStatus = typeof ListWithdrawalsStatus[keyof typeof ListWithdrawalsStatus];


export const ListWithdrawalsStatus = {
  pending: 'pending',
  approved: 'approved',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export type GetLeaderboardParams = {
limit?: number;
};