# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telegram Bingo API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/telegram:
    post:
      operationId: telegramAuth
      tags: [auth]
      summary: Authenticate with Telegram init data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAuthInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid init data
  /auth/admin:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials

  # ─── USERS ──────────────────────────────────────────────────────────────────
...
  /deposits:
    get:
      operationId: listDeposits
      tags: [transactions]
      summary: List deposits (admin gets all, user gets own)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Paginated deposits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositPage"
    post:
      operationId: createDeposit
      tags: [transactions]
      summary: Submit a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /deposits/{id}/approve:
    patch:
      operationId: approveDeposit
      tags: [transactions]
      summary: Approve a deposit (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approved deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /deposits/{id}/reject:
    patch:
      operationId: rejectDeposit
      tags: [transactions]
      summary: Reject a deposit (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"

  # ─── WITHDRAWALS ────────────────────────────────────────────────────────────
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── DASHBOARD ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TelegramAuthInput:
      type: object
      required: [initData]
      properties:
        initData: { type: string }
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password: { type: string }
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"

    # ─── USER ─────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, telegramId, username, firstName, balance, role, isBanned, createdAt]
      properties:
        id: { type: integer }
        telegramId: { type: string }
        username: { type: string }
        firstName: { type: string }
        lastName: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        balance: { type: number }
        totalWinnings: { type: number }
        totalGames: { type: integer }
        totalWins: { type: integer }
        role: { type: string, enum: [player, admin] }
        isBanned: { type: boolean }
        referralCode: { type: string }
        referredBy: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
...
    Deposit:
      type: object
      required: [id, userId, amount, status, telebirrNumber, transactionId, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        username: { type: ["string", "null"] }
        amount: { type: number }
        status: { type: string, enum: [pending, approved, rejected] }
        telebirrNumber: { type: string }
        transactionId: { type: string }
        screenshotUrl: { type: ["string", "null"] }
        rejectionReason: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    DepositPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Deposit"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    DepositInput:
      type: object
      required: [amount, telebirrNumber, transactionId]
      properties:
        amount: { type: number, minimum: 1 }
        telebirrNumber: { type: string }
        transactionId: { type: string }
        screenshotUrl: { type: string }

    # ─── WITHDRAWAL ───────────────────────────────────────────────────────────
...
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string }

    # ─── LEADERBOARD ──────────────────────────────────────────────────────────
...
    Announcement:
      type: object
      required: [id, title, message, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        message: { type: string }
        createdAt: { type: string, format: date-time }
    AnnouncementInput:
      type: object
      required: [title, message]
      properties:
        title: { type: string }
        message: { type: string }

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
...